[REFACTOR] prevent the user from accessing the implementations of classes
When origininally deciding to make the implementations accessible, testing if it works was done by creating a window using the SDL API in the client app. The reason it worked is because the client app was linked to FrongeEntry, which is a static library that links to SDL, thus making that window creation possible, but other stuff (like SDL_ttf) will not work.